+++ src/main/java/es/upm/grise/profundizacion2018/tema5/Document.java
@@ -13,10 +13,14 @@ public class Document {
 	private String title;
 	private String body;
 	
-	public Document() throws NonRecoverableError {
-		this.documentId = DocumentIdProvider.getInstance().getDocumentId();
+//	public Document() throws NonRecoverableError {
+//		this.documentId = DocumentIdProvider.getInstance().getDocumentId();
+//	}
+	
+	public void setDocId(DocumentIdProvider doc) throws NonRecoverableError{
+		this.documentId = DocumentIdProvider.getInstance(DocumentIdProvider.DRIVER).getDocumentId();
 	}
-
+	
 	public void setTemplate(String template) {
 		this.template = template;
 	}