+++ src/main/java/es/upm/grise/profundizacion2018/tema5/HandleDocuments.java
@@ -16,6 +16,7 @@ public class HandleDocuments {
 		try {
 			
 			Document document = new Document();
+			document.setDocId(DocumentIdProvider.getInstance(DocumentIdProvider.DRIVER));
 			document.setTemplate(TEMPLATE);
 			document.setAuthor(AUTHOR);
 			document.setTitle(TITLE);